Ensure error handler can access OutboundMessageRequest
